feat: add farcaster setup dialog
A new component, FarcasterSetupDialog, has been implemented to be used for managing account key signing in the Camp application. The FarcasterSetupDialog component can be used to fetch Farcaster account keys, post new keys, perform long-polling for updated key data, and display the relevant information in a user-friendly format. The component has been added to the application through the client-app-provider.js file which manages the loading of this type of component.