feat: add Farcaster account key route
This commit adds a new farcaster-account-key route. Through this new route, signer key pairs are generated and a signed key request is created using campAccount. Public key and status information can be received using the GET method, and the POST method is intended for farcaster signer key requests. The expiration of the keys is set to 2 hours.